[ENHANCEMENT] Allow Retries With Curl **change severity: moderate**
The change concerns the crucial part of sending notifications to slack; it now uses the '--retry' parameter of curl.
A maximum of 8 retries are made, and the delay between the first is 1 seconds and then with every retry it is doubled. For 8 retries 8.51 seconds will be elapsed, enough for any server errors to get rectified.

NOTE: [Only FTP 4xx and HTTP 5xx](https://stackoverflow.com/questions/42873285/curl-retry-mechanism/49825264#49825264) errors are caught and retried for.