Add slack update hook for new article feedback
Patch ArticleFeedbackv5 to add a new 'aftv5-newLogEntry' mediawiki
hook, and listen to it in SlackNotifications to send a slack update.

This approach is brittle because if the patch fails to apply on future
versions of AFTv5, the build will break, and will need to be manually
tweaked. Still, this seems to be the path of least resistance for now.